IImplement rules for version checking on specific modules
