+++ library/PackageUpdateInfo/PackageUpdateInfo/PackageUpdateInfo.csproj
@@ -11,8 +11,9 @@
   
   <PropertyGroup>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-    <AssemblyVersion>1.0.2.0</AssemblyVersion>
+    <AssemblyVersion>1.2.1.0</AssemblyVersion>
     <Version>1.2.0</Version>
+    <FileVersion>1.2.1.0</FileVersion>
   </PropertyGroup>
 
 </Project>
